+++ src/librustc_codegen_ssa/Cargo.toml
@@ -20,6 +20,7 @@ log = "0.4.5"
 libc = "0.2.44"
 jobserver = "0.1.11"
 parking_lot = "0.7"
+tempfile = "3.0.5"
 
 serialize = { path = "../libserialize" }
 syntax = { path = "../libsyntax" }
