+++ src/librustc_codegen_ssa/back/mod.rs
@@ -5,3 +5,5 @@ pub mod link;
 pub mod command;
 pub mod symbol_export;
 pub mod archive;
+pub mod rpath;
+pub mod wasm;
